@@ -21,7 +21,7 @@ namespace Microsoft.BotBuilderSamples.SimpleRootBot
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddControllers().AddNewtonsoftJson();
 
             // Configure credentials
             services.AddSingleton<ICredentialProvider, ConfigurationCredentialProvider>();
@@ -53,20 +53,19 @@ namespace Microsoft.BotBuilderSamples.SimpleRootBot
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseHsts();
-            }
+            app.UseDefaultFiles()
+                            .UseStaticFiles()
+                            .UseWebSockets()
+                            .UseRouting()
+                            .UseAuthorization()
+                            .UseEndpoints(endpoints =>
+                            {
+                                endpoints.MapControllers();
+                            });
 
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
-            app.UseMvc();
+            // app.UseHttpsRedirection();
         }
     }
 }
