@@ -16,7 +16,7 @@ namespace Microsoft.BotBuilderSamples.SimpleRootBot.Authentication
     /// </summary>
     public class AllowedSkillsClaimsValidator : ClaimsValidator
     {
-        private readonly List<string> _allowedSkills;
+        private readonly List<string> _allowedCallers;
 
         public AllowedSkillsClaimsValidator(SkillsConfiguration skillsConfig)
         {
@@ -26,18 +26,19 @@ namespace Microsoft.BotBuilderSamples.SimpleRootBot.Authentication
             }
 
             // Load the appIds for the configured skills (we will only allow responses from skills we have configured).
-            _allowedSkills = (from skill in skillsConfig.Skills.Values select skill.AppId).ToList();
+            _allowedCallers = (from skill in skillsConfig.Skills.Values select skill.AppId.ToUpperInvariant()).ToList();
         }
 
         public override Task ValidateClaimsAsync(IList<Claim> claims)
         {
-            if (SkillValidation.IsSkillClaim(claims))
+            // if _allowedCallers has one item of '*', allow all skill calls and do not validate the appid from claims
+            if (SkillValidation.IsSkillClaim(claims) && !(_allowedCallers.Count == 1 && _allowedCallers[0] == "*"))
             {
                 // Check that the appId claim in the skill request is in the list of skills configured for this bot.
-                var appId = JwtTokenValidation.GetAppIdFromClaims(claims);
-                if (!_allowedSkills.Contains(appId))
+                var appId = JwtTokenValidation.GetAppIdFromClaims(claims).ToUpperInvariant();
+                if (!_allowedCallers.Contains(appId))
                 {
-                    throw new UnauthorizedAccessException($"Received a request from an application with an appID of \"{appId}\". To enable requests from this skill, add the skill to your configuration file.");
+                    throw new UnauthorizedAccessException($"Received a request from a bot with an app ID of \"{appId}\". To enable requests from this caller, add the app ID to your configuration file.");
                 }
             }
 
